[refactor] base entity 필드 추가 및 id 타입 변경에 따른 수정 (#44)
* refactor: BaseEntity 개선 및 코드 정리

- id 타입을 int에서 long으로 변경
- @MappedSuperclass가 있으므로 abstract 선언 불필요
- Lombok 제거
- 필드의 접근 제어자 (`protected set`)를 제거

* refactor: BaseEntity id 타입을 int에서 long으로 변경에 따른 수정

* chore: 불필요한  @Builder.Default 어노테이션 제거

* refactor: lateinit var 로 수정 후 null 정리

+++ backend/src/main/java/com/bookbook/domain/chat/controller/WebSocketChatController.java
@@ -35,7 +35,7 @@ public class WebSocketChatController {
             }
             
             CustomOAuth2User user = (CustomOAuth2User) auth.getPrincipal();
-            Integer userId = user.getUserId().intValue();
+            Long userId = user.getUserId();
             
             log.info("WebSocket 메시지 전송 - roomId: {}, userId: {}", request.getRoomId(), userId);
             
@@ -94,7 +94,7 @@ public class WebSocketChatController {
             }
             
             CustomOAuth2User user = (CustomOAuth2User) auth.getPrincipal();
-            Integer userId = user.getUserId().intValue();
+            Long userId = user.getUserId();
             
             log.info("WebSocket 읽음 처리 - roomId: {}, userId: {}", roomId, userId);
